+++ apps/laboratory/src/pages/library/ethers-all.tsx
@@ -24,9 +24,7 @@ const modal = createWeb3Modal({
   termsConditionsUrl: 'https://walletconnect.com/terms',
   privacyPolicyUrl: 'https://walletconnect.com/privacy',
   siweConfig,
-  enableOnramp: true,
-  customWallets: ConstantsUtil.CustomWallets,
-  enableWalletFeatures: true
+  customWallets: ConstantsUtil.CustomWallets
 })
 
 ThemeStore.setModal(modal)