+++ packages/core/src/controllers/EnsController.ts
@@ -109,10 +109,6 @@ export const EnsController = {
       throw new Error('Address or auth connector not found')
     }
 
-    if (!this.isAllowedToRegisterName()) {
-      throw new Error('Not allowed to register name')
-    }
-
     state.loading = true
 
     try {
@@ -163,12 +159,5 @@ export const EnsController = {
     const ensError = error as BlockchainApiEnsError
 
     return ensError?.reasons?.[0]?.description || defaultError
-  },
-  isAllowedToRegisterName() {
-    const emailConnector = ConnectorController.getAuthConnector()
-    const email = emailConnector?.provider.getEmail() || ''
-    const domain = email.split('@')?.[1]
-
-    return domain && ConstantsUtil.WC_NAMES_ALLOWED_DOMAINS.includes(domain)
   }
 }