+++ packages/core/src/utils/RouterUtil.ts
@@ -1,7 +1,6 @@
 import { RouterController } from '../controllers/RouterController.js'
 import { ModalController } from '../controllers/ModalController.js'
 import { OptionsController } from '../controllers/OptionsController.js'
-import { AccountController } from '../controllers/AccountController.js'
 import { ChainController } from '../controllers/ChainController.js'
 
 export const RouterUtil = {
@@ -23,17 +22,14 @@ export const RouterUtil = {
   },
   navigateAfterPreferredAccountTypeSelect() {
     const { isSiweEnabled } = OptionsController.state
-    const profileName = AccountController.getProperty('profileName')
     if (isSiweEnabled && ChainController.state.activeChain === 'evm') {
       console.log(
         '>>> [RouterUtil] navigateAfterPreferredAccountTypeSelect: ConnectingSiwe',
         ChainController.state.activeChain
       )
       RouterController.push('ConnectingSiwe')
-    } else if (profileName) {
-      RouterController.push('Account')
     } else {
-      RouterController.push('ChooseAccountName')
+      RouterController.push('Account')
     }
   }
 }