+++ packages/core/src/utils/TypeUtil.ts
@@ -1,9 +1,14 @@
-import type { W3mFrameProvider } from '@web3modal/wallet'
+import type { W3mFrameProvider, W3mFrameTypes } from '@web3modal/wallet'
 import type { Balance, Transaction } from '@web3modal/common'
-import type { NetworkControllerClient } from '../controllers/NetworkController'
+import type {
+  NetworkControllerClient,
+  NetworkControllerState
+} from '../controllers/NetworkController'
 import type { ConnectionControllerClient } from '../controllers/ConnectionController'
 import type { OptionsControllerState } from '../controllers/OptionsController'
 import type { Chain } from '../controllers/ChainController'
+import type { AccountControllerState } from '../controllers/AccountController'
+import type { OnRampProviderOption } from '../controllers/OnRampController.js'
 
 export type CaipAddress = `${string}:${string}:${string}`
 
@@ -63,6 +68,7 @@ export type Connector = {
   socials?: SocialProvider[]
   showWallets?: boolean
   chain: Chain
+  walletFeatures?: boolean
 }
 
 export interface AuthConnector extends Connector {
@@ -437,6 +443,9 @@ export type Event =
   | {
       type: 'track'
       event: 'CLICK_TRANSACTIONS'
+      properties: {
+        isSmartAccount: boolean
+      }
     }
   | {
       type: 'track'
@@ -445,6 +454,7 @@ export type Event =
         address: string
         projectId: string
         cursor: string | undefined
+        isSmartAccount: boolean
       }
     }
   | {
@@ -454,15 +464,24 @@ export type Event =
         address: string | undefined
         projectId: string
         cursor: string | undefined
+        isSmartAccount: boolean
       }
     }
   | {
       type: 'track'
       event: 'CLICK_SIGN_SIWE_MESSAGE'
+      properties: {
+        network: string
+        isSmartAccount: boolean
+      }
     }
   | {
       type: 'track'
       event: 'CLICK_CANCEL_SIWE'
+      properties: {
+        network: string
+        isSmartAccount: boolean
+      }
     }
   | {
       type: 'track'
@@ -471,10 +490,18 @@ export type Event =
   | {
       type: 'track'
       event: 'SIWE_AUTH_SUCCESS'
+      properties: {
+        network: string
+        isSmartAccount: boolean
+      }
     }
   | {
       type: 'track'
       event: 'SIWE_AUTH_ERROR'
+      properties: {
+        network: string
+        isSmartAccount: boolean
+      }
     }
   | {
       type: 'track'
@@ -527,7 +554,171 @@ export type Event =
       type: 'track'
       event: 'CLICK_SELECT_NETWORK_TO_SWAP'
     }
-
+  | {
+      type: 'track'
+      event: 'SELECT_BUY_CRYPTO'
+      properties: {
+        isSmartAccount: boolean
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SELECT_BUY_PROVIDER'
+      properties: {
+        provider: OnRampProviderOption
+        isSmartAccount: boolean
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SELECT_WHAT_IS_A_BUY'
+      properties: {
+        isSmartAccount: boolean
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SET_PREFERRED_ACCOUNT_TYPE'
+      properties: {
+        accountType: W3mFrameTypes.AccountType
+        network: string
+      }
+    }
+  | {
+      type: 'track'
+      event: 'OPEN_SWAP'
+      properties: {
+        isSmartAccount: boolean
+        network: string
+      }
+    }
+  | {
+      type: 'track'
+      event: 'INITIATE_SWAP'
+      properties: {
+        isSmartAccount: boolean
+        network: string
+        swapFromToken: string
+        swapToToken: string
+        swapfromAmount: string
+        swapToAmount: string
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SWAP_SUCCESS'
+      properties: {
+        isSmartAccount: boolean
+        network: string
+        swapFromToken: string
+        swapToToken: string
+        swapfromAmount: string
+        swapToAmount: string
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SWAP_ERROR'
+      properties: {
+        isSmartAccount: boolean
+        network: string
+        swapFromToken: string
+        swapToToken: string
+        swapfromAmount: string
+        swapToAmount: string
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SOCIAL_LOGIN_STARTED'
+      properties: {
+        provider: SocialProvider
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SOCIAL_LOGIN_SUCCESS'
+      properties: {
+        provider: SocialProvider
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SOCIAL_LOGIN_ERROR'
+      properties: {
+        provider: SocialProvider
+      }
+    }
+  | {
+      type: 'track'
+      event: 'OPEN_ENS_FLOW'
+      properties: {
+        isSmartAccount: boolean
+      }
+    }
+  | {
+      type: 'track'
+      event: 'REGISTER_NAME_INITIATED'
+      properties: {
+        isSmartAccount: boolean
+        ensName: string
+      }
+    }
+  | {
+      type: 'track'
+      event: 'REGISTER_NAME_SUCCESS'
+      properties: {
+        isSmartAccount: boolean
+        ensName: string
+      }
+    }
+  | {
+      type: 'track'
+      event: 'REGISTER_NAME_ERROR'
+      properties: {
+        isSmartAccount: boolean
+        ensName: string
+        error: string
+      }
+    }
+  | {
+      type: 'track'
+      event: 'OPEN_SEND'
+      properties: {
+        isSmartAccount: boolean
+        network: string
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SEND_INITIATED'
+      properties: {
+        isSmartAccount: boolean
+        network: string
+        token: string
+        amount: number
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SEND_SUCCESS'
+      properties: {
+        isSmartAccount: boolean
+        network: string
+        token: string
+        amount: number
+      }
+    }
+  | {
+      type: 'track'
+      event: 'SEND_ERROR'
+      properties: {
+        isSmartAccount: boolean
+        network: string
+        token: string
+        amount: number
+      }
+    }
 // Onramp Types
 export type DestinationWallet = {
   address: string
@@ -614,10 +805,11 @@ export interface WriteContractArgs {
   abi: any
 }
 
-export interface AdapterCore {
-  protocol: 'evm' | 'solana'
-  networkControllerClient: NetworkControllerClient
-  connectionControllerClient: ConnectionControllerClient
-  construct(scaffold: any, options: OptionsControllerState): void
-  initialize(): void
+export type ChainAdapter = Partial<NetworkControllerState> & {
+  connectionControllerClient?: ConnectionControllerClient
+  networkControllerClient?: NetworkControllerClient
+  accountState: AccountControllerState
+  chain: Chain
+  construct?: (scaffold: any, options: OptionsControllerState) => void
+  initialize?: () => void
 }